fix: Renaming package for errors
